+++ packages/redux-routine/src/test/index.js
@@ -7,7 +7,6 @@ import { createStore, applyMiddleware } from 'redux';
  * Internal dependencies
  */
 import createMiddleware from '../middleware';
-import createRuntime from '../runtime';
 
 describe( 'createMiddleware', () => {
 	function createStoreWithMiddleware( middleware ) {
@@ -16,8 +15,7 @@ describe( 'createMiddleware', () => {
 	}
 
 	it( 'should not alter dispatch flow of uncontrolled action', () => {
-		const runtime = createRuntime();
-		const middleware = createMiddleware( runtime );
+		const middleware = createMiddleware();
 		const store = createStoreWithMiddleware( middleware );
 
 		store.dispatch( { type: 'CHANGE', nextState: 1 } );
@@ -26,8 +24,7 @@ describe( 'createMiddleware', () => {
 	} );
 
 	it( 'should dispatch yielded actions', () => {
-		const runtime = createRuntime();
-		const middleware = createMiddleware( runtime );
+		const middleware = createMiddleware();
 		const store = createStoreWithMiddleware( middleware );
 		function* createAction() {
 			yield { type: 'CHANGE', nextState: 1 };
@@ -38,78 +35,60 @@ describe( 'createMiddleware', () => {
 		expect( store.getState() ).toBe( 1 );
 	} );
 
-	it( 'should continue only once control condition resolves', ( done ) => {
-		const runtime = createRuntime( {
-			WAIT: () => new Promise( ( resolve ) => setTimeout( resolve, 0 ) ),
+	it( 'should continue only once control condition resolves', async () => {
+		const middleware = createMiddleware( {
+			WAIT: () => new Promise( ( resolve ) => resolve() ),
 		} );
-		const middleware = createMiddleware( runtime );
 		const store = createStoreWithMiddleware( middleware );
 		function* createAction() {
 			yield { type: 'WAIT' };
 			yield { type: 'CHANGE', nextState: 1 };
 		}
 
-		store.dispatch( createAction() );
-		expect( store.getState() ).toBe( null );
-
-		jest.runAllTimers();
-
-		// Promise resolution occurs on next tick.
-		process.nextTick( () => {
-			expect( store.getState() ).toBe( 1 );
-			done();
-		} );
+		await store.dispatch( createAction() );
+		expect( store.getState() ).toBe( 1 );
 	} );
 
-	it( 'should throw if promise rejects', ( done ) => {
-		const runtime = createRuntime( {
-			WAIT_FAIL: () => new Promise( ( resolve, reject ) => {
-				setTimeout( () => reject( 'Message' ), 0 );
-			} ),
+	it( 'should throw if promise rejects', async () => {
+		const middleware = createMiddleware( {
+			WAIT_FAIL: () => new Promise( ( resolve, reject ) =>
+				reject( 'Message' )
+			),
 		} );
-		const middleware = createMiddleware( runtime );
 		const store = createStoreWithMiddleware( middleware );
 		function* createAction() {
 			try {
 				yield { type: 'WAIT_FAIL' };
 			} catch ( error ) {
 				expect( error.message ).toBe( 'Message' );
-				done();
 			}
 		}
 
-		store.dispatch( createAction() );
-
-		jest.runAllTimers();
+		await store.dispatch( createAction() );
 	} );
 
-	it( 'should throw if promise throws', ( done ) => {
-		const runtime = createRuntime( {
+	it( 'should throw if promise throws', () => {
+		const middleware = createMiddleware( {
 			WAIT_FAIL: () => new Promise( () => {
 				throw new Error( 'Message' );
 			} ),
 		} );
-		const middleware = createMiddleware( runtime );
 		const store = createStoreWithMiddleware( middleware );
 		function* createAction() {
 			try {
 				yield { type: 'WAIT_FAIL' };
 			} catch ( error ) {
 				expect( error.message ).toBe( 'Message' );
-				done();
 			}
 		}
 
-		store.dispatch( createAction() );
-
-		jest.runAllTimers();
+		return store.dispatch( createAction() );
 	} );
 
 	it( 'assigns sync controlled return value into yield assignment', () => {
-		const runtime = createRuntime( {
+		const middleware = createMiddleware( {
 			RETURN_TWO: () => 2,
 		} );
-		const middleware = createMiddleware( runtime );
 		const store = createStoreWithMiddleware( middleware );
 		function* createAction() {
 			const nextState = yield { type: 'RETURN_TWO' };
@@ -121,45 +100,20 @@ describe( 'createMiddleware', () => {
 		expect( store.getState() ).toBe( 2 );
 	} );
 
-	it( 'assigns async controlled return value into yield assignment', ( done ) => {
-		const runtime = createRuntime( {
+	it( 'assigns async controlled return value into yield assignment', async () => {
+		const middleware = createMiddleware( {
 			WAIT: ( action ) => new Promise( ( resolve ) => {
-				setTimeout( () => {
-					resolve( action.value );
-				}, 0 );
+				resolve( action.value );
 			} ),
 		} );
-		const middleware = createMiddleware( runtime );
 		const store = createStoreWithMiddleware( middleware );
 		function* createAction() {
 			const nextState = yield { type: 'WAIT', value: 2 };
 			return { type: 'CHANGE', nextState };
 		}
 
-		store.dispatch( createAction() );
-		expect( store.getState() ).toBe( null );
+		await store.dispatch( createAction() );
 
-		jest.runAllTimers();
-
-		process.nextTick( () => {
-			expect( store.getState() ).toBe( 2 );
-			done();
-		} );
-	} );
-
-	it( 'kills continuation if control returns undefined', () => {
-		const runtime = createRuntime( {
-			KILL: () => { },
-		} );
-		const middleware = createMiddleware( runtime );
-		const store = createStoreWithMiddleware( middleware );
-		function* createAction() {
-			yield { type: 'KILL' };
-			return { type: 'CHANGE', nextState: 1 };
-		}
-
-		store.dispatch( createAction() );
-
-		expect( store.getState() ).toBe( null );
+		expect( store.getState() ).toBe( 2 );
 	} );
 } );