+++ packages/redux-routine/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Breaking change
 
-- This releases changes the public API of the package. We're now exporting `createRuntime` and `createMiddleware` APIs to allow more flexibility in using the co-routine runtime without using it as a Redux Middleware.
+- The middleware returns a promise resolving once the runtime finishes looping throught the generator.
+- It's not possible to kill the execution of the runtime anymore by returning `undefined`
 
 ## 2.0.0 (2018-09-05)